created request-types.ts to unify data structure
